server/requirements.txt
Django==5.1.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
psycopg[pool]
python-dotenv==1.0.1
pytest
pytest-django
django-cors-headers


